@@ -1,7 +1,6 @@
 name: flushbar
 description: A flexible widget for user notification. Customize your text, button, duration, animations and much more. For Android devs, it is made to replace Snackbars and Toasts.
-version: 1.7.0
-author: Andre Haueisen <andrehaueisen@gmail.com>
+version: 1.10.2
 homepage: https://github.com/AndreHaueisen/flushbar
 
 dependencies:
@@ -13,11 +12,13 @@ dev_dependencies:
     sdk: flutter
 
 environment:
-  sdk: ">=2.0.0-dev.52.0 <3.0.0"
-  flutter: ">=0.4.0"
+  sdk: ">=2.7.2 <3.0.0"
+  flutter: ">=1.5.0 <2.0.0"
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
+
+# publish command: flutter packages pub publish --dry-run
 
 # The following section is specific to Flutter.
 flutter:
