@@ -1,32 +1,41 @@
-import 'package:flushbar/flushbar.dart';
-import 'package:flutter/material.dart';
-
 import 'dart:async';
 import 'dart:ui';
 
+import 'package:flushbar/flushbar.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 
 class FlushbarRoute<T> extends OverlayRoute<T> {
+  final Flushbar flushbar;
+  final Builder _builder;
+  final Completer<T> _transitionCompleter = Completer<T>();
+  final FlushbarStatusCallback _onStatusChanged;
+
+  Animation<double> _filterBlurAnimation;
+  Animation<Color> _filterColorAnimation;
+  Alignment _initialAlignment;
+  Alignment _endAlignment;
+  bool _wasDismissedBySwipe = false;
+  Timer _timer;
+  T _result;
+  FlushbarStatus currentStatus;
+
   FlushbarRoute({
-    @required this.theme,
     @required this.flushbar,
     RouteSettings settings,
-  }) : super(settings: settings) {
-    this._builder = Builder(builder: (BuildContext innerContext) {
-      return GestureDetector(
-        child: flushbar,
-        onTap: flushbar.onTap != null ? () {
-          flushbar.onTap(flushbar);
-        } : null,
-      );
-    });
-
+  })  : _builder = Builder(builder: (BuildContext innerContext) {
+          return GestureDetector(
+            child: flushbar,
+            onTap:
+                flushbar.onTap != null ? () => flushbar.onTap(flushbar) : null,
+          );
+        }),
+        _onStatusChanged = flushbar.onStatusChanged,
+        super(settings: settings) {
     _configureAlignment(this.flushbar.flushbarPosition);
-
-    _onStatusChanged = flushbar.onStatusChanged;
   }
 
-  _configureAlignment(FlushbarPosition flushbarPosition) {
+  void _configureAlignment(FlushbarPosition flushbarPosition) {
     switch (flushbar.flushbarPosition) {
       case FlushbarPosition.TOP:
         {
@@ -43,37 +52,20 @@ class FlushbarRoute<T> extends OverlayRoute<T> {
     }
   }
 
-  Flushbar flushbar;
-  Builder _builder;
-
-  final ThemeData theme;
-
   Future<T> get completed => _transitionCompleter.future;
-  final Completer<T> _transitionCompleter = Completer<T>();
-
-  FlushbarStatusCallback _onStatusChanged;
-  Alignment _initialAlignment;
-  Alignment _endAlignment;
-  bool _wasDismissedBySwipe = false;
-
-  Timer _timer;
-
   bool get opaque => false;
 
   @override
   Iterable<OverlayEntry> createOverlayEntries() {
-    List<OverlayEntry> overlays = [];
+    final List<OverlayEntry> overlays = [];
 
-    if (flushbar.overlayBlur > 0.0) {
+    if (flushbar.blockBackgroundInteraction) {
       overlays.add(
         OverlayEntry(
             builder: (BuildContext context) {
-              return BackdropFilter(
-                filter: ImageFilter.blur(sigmaX: flushbar.overlayBlur, sigmaY: flushbar.overlayBlur),
-                child: Container(
-                  constraints: BoxConstraints.expand(),
-                  color: flushbar.overlayColor,
-                ),
+              return GestureDetector(
+                onTap: flushbar.isDismissible ? () => flushbar.dismiss() : null,
+                child: _createBackgroundOverlay(),
               );
             },
             maintainState: false,
@@ -87,20 +79,74 @@ class FlushbarRoute<T> extends OverlayRoute<T> {
             final Widget annotatedChild = Semantics(
               child: AlignTransition(
                 alignment: _animation,
-                child:
-                    flushbar.isDismissible ? _getDismissibleFlushbar(_builder) : Padding(padding: flushbar.aroundPadding, child: _builder),
+                child: flushbar.isDismissible
+                    ? _getDismissibleFlushbar(_builder)
+                    : _getFlushbar(),
               ),
-              focused: true,
-              scopesRoute: true,
+              focused: false,
+              container: true,
               explicitChildNodes: true,
             );
-            return theme != null ? Theme(data: theme, child: annotatedChild) : annotatedChild;
+            return annotatedChild;
           },
           maintainState: false,
           opaque: opaque),
     );
 
     return overlays;
+  }
+
+  Widget _createBackgroundOverlay() {
+    if (_filterBlurAnimation != null && _filterColorAnimation != null) {
+      return AnimatedBuilder(
+        animation: _filterBlurAnimation,
+        builder: (context, child) {
+          return BackdropFilter(
+            filter: ImageFilter.blur(
+                sigmaX: _filterBlurAnimation.value,
+                sigmaY: _filterBlurAnimation.value),
+            child: Container(
+              constraints: BoxConstraints.expand(),
+              color: _filterColorAnimation.value,
+            ),
+          );
+        },
+      );
+    }
+
+    if (_filterBlurAnimation != null) {
+      return AnimatedBuilder(
+        animation: _filterBlurAnimation,
+        builder: (context, child) {
+          return BackdropFilter(
+            filter: ImageFilter.blur(
+                sigmaX: _filterBlurAnimation.value,
+                sigmaY: _filterBlurAnimation.value),
+            child: Container(
+              constraints: BoxConstraints.expand(),
+              color: Colors.transparent,
+            ),
+          );
+        },
+      );
+    }
+
+    if (_filterColorAnimation != null) {
+      AnimatedBuilder(
+        animation: _filterColorAnimation,
+        builder: (context, child) {
+          return Container(
+            constraints: BoxConstraints.expand(),
+            color: _filterColorAnimation.value,
+          );
+        },
+      );
+    }
+
+    return Container(
+      constraints: BoxConstraints.expand(),
+      color: Colors.transparent,
+    );
   }
 
   /// This string is a workaround until Dismissible supports a returning item
@@ -111,7 +157,8 @@ class FlushbarRoute<T> extends OverlayRoute<T> {
       direction: _getDismissDirection(),
       resizeDuration: null,
       confirmDismiss: (_) {
-        if (currentStatus == FlushbarStatus.IS_APPEARING || currentStatus == FlushbarStatus.IS_HIDING) {
+        if (currentStatus == FlushbarStatus.IS_APPEARING ||
+            currentStatus == FlushbarStatus.IS_HIDING) {
           return Future.value(false);
         }
         return Future.value(true);
@@ -128,10 +175,7 @@ class FlushbarRoute<T> extends OverlayRoute<T> {
           navigator.removeRoute(this);
         }
       },
-      child: Padding(
-        padding: flushbar.aroundPadding,
-        child: child,
-      ),
+      child: _getFlushbar(),
     );
   }
 
@@ -147,8 +191,16 @@ class FlushbarRoute<T> extends OverlayRoute<T> {
     }
   }
 
+  Widget _getFlushbar() {
+    return Container(
+      margin: flushbar.margin,
+      child: _builder,
+    );
+  }
+
   @override
-  bool get finishedWhenPopped => _controller.status == AnimationStatus.dismissed;
+  bool get finishedWhenPopped =>
+      _controller.status == AnimationStatus.dismissed;
 
   /// The animation that drives the route's transition and the previous route's
   /// forward transition.
@@ -166,8 +218,10 @@ class FlushbarRoute<T> extends OverlayRoute<T> {
   /// this route from the previous one, and back to the previous route from this
   /// one.
   AnimationController createAnimationController() {
-    assert(!_transitionCompleter.isCompleted, 'Cannot reuse a $runtimeType after disposing it.');
-    assert(flushbar.animationDuration != null && flushbar.animationDuration >= Duration.zero);
+    assert(!_transitionCompleter.isCompleted,
+        'Cannot reuse a $runtimeType after disposing it.');
+    assert(flushbar.animationDuration != null &&
+        flushbar.animationDuration >= Duration.zero);
     return AnimationController(
       duration: flushbar.animationDuration,
       debugLabel: debugLabel,
@@ -179,7 +233,8 @@ class FlushbarRoute<T> extends OverlayRoute<T> {
   /// the transition controlled by the animation controller created by
   /// [createAnimationController()].
   Animation<Alignment> createAnimation() {
-    assert(!_transitionCompleter.isCompleted, 'Cannot reuse a $runtimeType after disposing it.');
+    assert(!_transitionCompleter.isCompleted,
+        'Cannot reuse a $runtimeType after disposing it.');
     assert(_controller != null);
     return AlignmentTween(begin: _initialAlignment, end: _endAlignment).animate(
       CurvedAnimation(
@@ -190,8 +245,36 @@ class FlushbarRoute<T> extends OverlayRoute<T> {
     );
   }
 
-  T _result;
-  FlushbarStatus currentStatus;
+  Animation<double> createBlurFilterAnimation() {
+    if (flushbar.routeBlur == null) return null;
+
+    return Tween(begin: 0.0, end: flushbar.routeBlur).animate(
+      CurvedAnimation(
+        parent: _controller,
+        curve: Interval(
+          0.0,
+          0.35,
+          curve: Curves.easeInOutCirc,
+        ),
+      ),
+    );
+  }
+
+  Animation<Color> createColorFilterAnimation() {
+    if (flushbar.routeColor == null) return null;
+
+    return ColorTween(begin: Colors.transparent, end: flushbar.routeColor)
+        .animate(
+      CurvedAnimation(
+        parent: _controller,
+        curve: Interval(
+          0.0,
+          0.35,
+          curve: Curves.easeInOutCirc,
+        ),
+      ),
+    );
+  }
 
   //copy of `routes.dart`
   void _handleStatusChanged(AnimationStatus status) {
@@ -230,37 +313,49 @@ class FlushbarRoute<T> extends OverlayRoute<T> {
   }
 
   @override
-  void install(OverlayEntry insertionPoint) {
-    assert(!_transitionCompleter.isCompleted, 'Cannot install a $runtimeType after disposing it.');
+  void install() {
+    assert(!_transitionCompleter.isCompleted,
+        'Cannot install a $runtimeType after disposing it.');
     _controller = createAnimationController();
-    assert(_controller != null, '$runtimeType.createAnimationController() returned null.');
+    assert(_controller != null,
+        '$runtimeType.createAnimationController() returned null.');
+    _filterBlurAnimation = createBlurFilterAnimation();
+    _filterColorAnimation = createColorFilterAnimation();
     _animation = createAnimation();
     assert(_animation != null, '$runtimeType.createAnimation() returned null.');
-    super.install(insertionPoint);
+    super.install();
   }
 
   @override
   TickerFuture didPush() {
-    assert(_controller != null, '$runtimeType.didPush called before calling install() or after calling dispose().');
-    assert(!_transitionCompleter.isCompleted, 'Cannot reuse a $runtimeType after disposing it.');
+    assert(_controller != null,
+        '$runtimeType.didPush called before calling install() or after calling dispose().');
+    assert(!_transitionCompleter.isCompleted,
+        'Cannot reuse a $runtimeType after disposing it.');
     _animation.addStatusListener(_handleStatusChanged);
     _configureTimer();
+    super.didPush();
     return _controller.forward();
   }
 
   @override
   void didReplace(Route<dynamic> oldRoute) {
-    assert(_controller != null, '$runtimeType.didReplace called before calling install() or after calling dispose().');
-    assert(!_transitionCompleter.isCompleted, 'Cannot reuse a $runtimeType after disposing it.');
-    if (oldRoute is FlushbarRoute) _controller.value = oldRoute._controller.value;
+    assert(_controller != null,
+        '$runtimeType.didReplace called before calling install() or after calling dispose().');
+    assert(!_transitionCompleter.isCompleted,
+        'Cannot reuse a $runtimeType after disposing it.');
+    if (oldRoute is FlushbarRoute)
+      _controller.value = oldRoute._controller.value;
     _animation.addStatusListener(_handleStatusChanged);
     super.didReplace(oldRoute);
   }
 
   @override
   bool didPop(T result) {
-    assert(_controller != null, '$runtimeType.didPop called before calling install() or after calling dispose().');
-    assert(!_transitionCompleter.isCompleted, 'Cannot reuse a $runtimeType after disposing it.');
+    assert(_controller != null,
+        '$runtimeType.didPop called before calling install() or after calling dispose().');
+    assert(!_transitionCompleter.isCompleted,
+        'Cannot reuse a $runtimeType after disposing it.');
 
     _result = result;
     _cancelTimer();
@@ -317,7 +412,8 @@ class FlushbarRoute<T> extends OverlayRoute<T> {
 
   @override
   void dispose() {
-    assert(!_transitionCompleter.isCompleted, 'Cannot dispose a $runtimeType twice.');
+    assert(!_transitionCompleter.isCompleted,
+        'Cannot dispose a $runtimeType twice.');
     _controller?.dispose();
     _transitionCompleter.complete(_result);
     super.dispose();
@@ -330,12 +426,12 @@ class FlushbarRoute<T> extends OverlayRoute<T> {
   String toString() => '$runtimeType(animation: $_controller)';
 }
 
-FlushbarRoute showFlushbar<T>({@required BuildContext context, @required Flushbar flushbar}) {
+FlushbarRoute showFlushbar<T>(
+    {@required BuildContext context, @required Flushbar flushbar}) {
   assert(flushbar != null);
 
   return FlushbarRoute<T>(
     flushbar: flushbar,
-    theme: Theme.of(context),
     settings: RouteSettings(name: FLUSHBAR_ROUTE_NAME),
   );
 }
