@@ -1,47 +1,17 @@
-import 'package:flutter/material.dart';
-
 import 'dart:async';
 import 'dart:ui';
 
+import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
+
 import 'flushbar_route.dart' as route;
 
 const String FLUSHBAR_ROUTE_NAME = "/flushbarRoute";
 
 typedef void FlushbarStatusCallback(FlushbarStatus status);
+typedef void OnTap(Flushbar flushbar);
 
 /// A highly customizable widget so you can notify your user when you fell like he needs a beautiful explanation.
-///
-/// [title] The title displayed to the user
-/// [message] The message displayed to the user.
-/// [titleText] Replaces [title]. Although this accepts a [Widget], it is meant to receive [Text] or [RichText]
-/// [messageText] Replaces [message]. Although this accepts a [Widget], it is meant to receive [Text] or  [RichText]
-/// [icon] You can use any widget here, but I recommend [Icon] or [Image] as indication of what kind of message you are displaying. Other widgets may break the layout
-/// [shouldIconPulse] An option to animate the icon (if present). Defaults to true.
-/// [aroundPadding] Adds a custom padding to Flushbar
-/// [borderRadius] Adds a radius to all corners of Flushbar. Best combined with [aroundPadding]. I do not recommend using it with [showProgressIndicator] or [leftBarIndicatorColor]
-/// [backgroundColor] Flushbar background color. Will be ignored if [backgroundGradient] is not null.
-/// [leftBarIndicatorColor] If not null, shows a left vertical bar to better indicate the humor of the notification. It is not possible to use it with a [Form] and I do not recommend using it with [LinearProgressIndicator].
-/// [boxShadows] The shadows generated by Flushbar. Leave it null if you don't want a shadow. You can use more than one if you feel the need. Check (this example)[https://github.com/flutter/flutter/blob/master/packages/flutter/lib/src/material/shadows.dart]
-/// [backgroundGradient] Flushbar background gradient. Makes [backgroundColor] be ignored.
-/// [mainButton] A [FlatButton] widget if you need an action from the user.
-/// [onTap] A callback that registers the user's click anywhere. An alternative to [mainButton]
-/// [duration] How long until Flushbar will hide itself (be dismissed). To make it indefinite, leave it null.
-/// [isDismissible] Determines if the user can swipe to dismiss the bar. It is recommended that you set [duration] != null if [isDismissible] == false. If the user swipes Flushbar to dismiss it no value will be returned.
-/// [dismissDirection] FlushbarDismissDirection.VERTICAL by default. Can also be [FlushbarDismissDirection.HORIZONTAL] in which case both left and right dismiss are allowed.
-/// [flushbarPosition] Flushbar can be based on [FlushbarPosition.TOP] or on [FlushbarPosition.BOTTOM] of your screen. [FlushbarPosition.BOTTOM] is the default.
-/// [flushbarStyle] Flushbar can be floating or be grounded to the edge of the screen. If grounded, I do not recommend using [aroundPadding] or [borderRadius]. [FlushbarStyle.FLOATING] is the default
-/// [forwardAnimationCurve] The [Curve] animation used when show() is called. [Curves.easeOut] is default.
-/// [reverseAnimationCurve] The [Curve] animation used when dismiss() is called. [Curves.fastOutSlowIn] is default.
-/// [animationDuration] Use it to speed up or slow down the animation duration
-/// [showProgressIndicator] true if you want to show a [LinearProgressIndicator].
-/// [progressIndicatorController] An optional [AnimationController] when you want to control the progress of your [LinearProgressIndicator].
-/// [progressIndicatorBackgroundColor] a [LinearProgressIndicator] configuration parameter.
-/// [progressIndicatorValueColor] a [LinearProgressIndicator] configuration parameter.
-/// [overlayBlur] Default is 0.0. If different than 0.0, creates a blurred overlay that prevents the user from interacting with the screen. The greater the value, the greater the blur.
-/// [overlayColor] Default is [Colors.transparent]. Only takes effect if [overlayBlur] > 0.0. Make sure you use a color with transparency here e.g. Colors.grey[600].withOpacity(0.2).
-/// [userInputForm] A [TextFormField] in case you want a simple user input. Every other widget is ignored if this is not null.
-/// [onStatusChanged] a callback for you to listen to the different Flushbar status
 class Flushbar<T extends Object> extends StatefulWidget {
   Flushbar(
       {Key key,
@@ -51,29 +21,36 @@ class Flushbar<T extends Object> extends StatefulWidget {
       Widget messageText,
       Widget icon,
       bool shouldIconPulse = true,
-      EdgeInsets aroundPadding = const EdgeInsets.all(0.0),
+      double maxWidth,
+      EdgeInsets margin = const EdgeInsets.all(0.0),
+      EdgeInsets padding = const EdgeInsets.all(16),
       double borderRadius = 0.0,
+      Color borderColor,
+      double borderWidth = 1.0,
       Color backgroundColor = const Color(0xFF303030),
       Color leftBarIndicatorColor,
       List<BoxShadow> boxShadows,
       Gradient backgroundGradient,
       FlatButton mainButton,
-      Function onTap(Flushbar flushbar),
+      OnTap onTap,
       Duration duration,
       bool isDismissible = true,
-      FlushbarDismissDirection dismissDirection = FlushbarDismissDirection.VERTICAL,
+      FlushbarDismissDirection dismissDirection =
+          FlushbarDismissDirection.VERTICAL,
       bool showProgressIndicator = false,
       AnimationController progressIndicatorController,
       Color progressIndicatorBackgroundColor,
       Animation<Color> progressIndicatorValueColor,
       FlushbarPosition flushbarPosition = FlushbarPosition.BOTTOM,
       FlushbarStyle flushbarStyle = FlushbarStyle.FLOATING,
-      Curve forwardAnimationCurve = Curves.easeOut,
-      Curve reverseAnimationCurve = Curves.fastOutSlowIn,
+      Curve forwardAnimationCurve = Curves.easeOutCirc,
+      Curve reverseAnimationCurve = Curves.easeOutCirc,
       Duration animationDuration = const Duration(seconds: 1),
       FlushbarStatusCallback onStatusChanged,
-      double overlayBlur = 0.0,
-      Color overlayColor = Colors.transparent,
+      double barBlur = 0.0,
+      bool blockBackgroundInteraction = false,
+      double routeBlur,
+      Color routeColor,
       Form userInputForm})
       : this.title = title,
         this.message = message,
@@ -81,8 +58,12 @@ class Flushbar<T extends Object> extends StatefulWidget {
         this.messageText = messageText,
         this.icon = icon,
         this.shouldIconPulse = shouldIconPulse,
-        this.aroundPadding = aroundPadding,
+        this.maxWidth = maxWidth,
+        this.margin = margin,
+        this.padding = padding,
         this.borderRadius = borderRadius,
+        this.borderColor = borderColor,
+        this.borderWidth = borderWidth,
         this.backgroundColor = backgroundColor,
         this.leftBarIndicatorColor = leftBarIndicatorColor,
         this.boxShadows = boxShadows,
@@ -94,50 +75,153 @@ class Flushbar<T extends Object> extends StatefulWidget {
         this.dismissDirection = dismissDirection,
         this.showProgressIndicator = showProgressIndicator,
         this.progressIndicatorController = progressIndicatorController,
-        this.progressIndicatorBackgroundColor = progressIndicatorBackgroundColor,
+        this.progressIndicatorBackgroundColor =
+            progressIndicatorBackgroundColor,
         this.progressIndicatorValueColor = progressIndicatorValueColor,
         this.flushbarPosition = flushbarPosition,
         this.flushbarStyle = flushbarStyle,
         this.forwardAnimationCurve = forwardAnimationCurve,
         this.reverseAnimationCurve = reverseAnimationCurve,
         this.animationDuration = animationDuration,
-        this.overlayBlur = overlayBlur,
-        this.overlayColor = overlayColor,
+        this.barBlur = barBlur,
+        this.blockBackgroundInteraction = blockBackgroundInteraction,
+        this.routeBlur = routeBlur,
+        this.routeColor = routeColor,
         this.userInputForm = userInputForm,
         super(key: key) {
     this.onStatusChanged = onStatusChanged ?? (status) {};
   }
 
+  /// A callback for you to listen to the different Flushbar status
   FlushbarStatusCallback onStatusChanged;
+
+  /// The title displayed to the user
   final String title;
+
+  /// The message displayed to the user.
   final String message;
+
+  /// Replaces [title]. Although this accepts a [Widget], it is meant to receive [Text] or [RichText]
   final Widget titleText;
+
+  /// Replaces [message]. Although this accepts a [Widget], it is meant to receive [Text] or  [RichText]
   final Widget messageText;
+
+  /// Will be ignored if [backgroundGradient] is not null
   final Color backgroundColor;
+
+  /// If not null, shows a left vertical bar to better indicate the humor of the notification.
+  /// It is not possible to use it with a [Form] and I do not recommend using it with [LinearProgressIndicator]
   final Color leftBarIndicatorColor;
+
+  /// [boxShadows] The shadows generated by Flushbar. Leave it null if you don't want a shadow.
+  /// You can use more than one if you feel the need.
+  /// Check (this example)[https://github.com/flutter/flutter/blob/master/packages/flutter/lib/src/material/shadows.dart]
   final List<BoxShadow> boxShadows;
+
+  /// Makes [backgroundColor] be ignored.
   final Gradient backgroundGradient;
+
+  /// You can use any widget here, but I recommend [Icon] or [Image] as indication of what kind
+  /// of message you are displaying. Other widgets may break the layout
   final Widget icon;
+
+  /// An option to animate the icon (if present). Defaults to true.
   final bool shouldIconPulse;
+
+  /// A [FlatButton] widget if you need an action from the user.
   final FlatButton mainButton;
-  final Function(Flushbar) onTap;
+
+  /// A callback that registers the user's click anywhere. An alternative to [mainButton]
+  final OnTap onTap;
+
+  /// How long until Flushbar will hide itself (be dismissed). To make it indefinite, leave it null.
   final Duration duration;
+
+  /// True if you want to show a [LinearProgressIndicator].
+  /// If [progressIndicatorController] is null, an infinite progress indicator will be shown
   final bool showProgressIndicator;
+
+  /// An optional [AnimationController] when you want to control the progress of your [LinearProgressIndicator].
+  /// You are responsible for controlling the progress
   final AnimationController progressIndicatorController;
+
+  /// A [LinearProgressIndicator] configuration parameter.
   final Color progressIndicatorBackgroundColor;
+
+  /// A [LinearProgressIndicator] configuration parameter.
   final Animation<Color> progressIndicatorValueColor;
+
+  /// Determines if the user can swipe or click the overlay (if [routeBlur] > 0) to dismiss.
+  /// It is recommended that you set [duration] != null if this is false.
+  /// If the user swipes to dismiss or clicks the overlay, no value will be returned.
   final bool isDismissible;
-  final EdgeInsets aroundPadding;
+
+  /// Used to limit Flushbar width (usually on large screens)
+  final double maxWidth;
+
+  /// Adds a custom margin to Flushbar
+  final EdgeInsets margin;
+
+  /// Adds a custom padding to Flushbar
+  /// The default follows material design guide line
+  final EdgeInsets padding;
+
+  /// Adds a radius to all corners of Flushbar. Best combined with [margin].
+  /// I do not recommend using it with [showProgressIndicator] or [leftBarIndicatorColor].
   final double borderRadius;
-  final Form userInputForm;
+
+  /// Adds a border to every side of Flushbar
+  /// I do not recommend using it with [showProgressIndicator] or [leftBarIndicatorColor].
+  final Color borderColor;
+
+  /// Changes the width of the border if [borderColor] is specified
+  final double borderWidth;
+
+  /// Flushbar can be based on [FlushbarPosition.TOP] or on [FlushbarPosition.BOTTOM] of your screen.
+  /// [FlushbarPosition.BOTTOM] is the default.
   final FlushbarPosition flushbarPosition;
+
+  /// [FlushbarDismissDirection.VERTICAL] by default.
+  /// Can also be [FlushbarDismissDirection.HORIZONTAL] in which case both left and right dismiss are allowed.
   final FlushbarDismissDirection dismissDirection;
+
+  /// Flushbar can be floating or be grounded to the edge of the screen.
+  /// If grounded, I do not recommend using [margin] or [borderRadius]. [FlushbarStyle.FLOATING] is the default
+  /// If grounded, I do not recommend using a [backgroundColor] with transparency or [barBlur]
   final FlushbarStyle flushbarStyle;
+
+  /// The [Curve] animation used when show() is called. [Curves.easeOut] is default
   final Curve forwardAnimationCurve;
+
+  /// The [Curve] animation used when dismiss() is called. [Curves.fastOutSlowIn] is default
   final Curve reverseAnimationCurve;
+
+  /// Use it to speed up or slow down the animation duration
   final Duration animationDuration;
-  final double overlayBlur;
-  final Color overlayColor;
+
+  /// Default is 0.0. If different than 0.0, blurs only Flushbar's background.
+  /// To take effect, make sure your [backgroundColor] has some opacity.
+  /// The greater the value, the greater the blur.
+  final double barBlur;
+
+  /// Determines if user can interact with the screen behind it
+  /// If this is false, [routeBlur] and [routeColor] will be ignored
+  final bool blockBackgroundInteraction;
+
+  /// Default is 0.0. If different than 0.0, creates a blurred
+  /// overlay that prevents the user from interacting with the screen.
+  /// The greater the value, the greater the blur.
+  /// It does not take effect if [blockBackgroundInteraction] is false
+  final double routeBlur;
+
+  /// Default is [Colors.transparent]. Only takes effect if [routeBlur] > 0.0.
+  /// Make sure you use a color with transparency here e.g. Colors.grey[600].withOpacity(0.2).
+  /// It does not take effect if [blockBackgroundInteraction] is false
+  final Color routeColor;
+
+  /// A [TextFormField] in case you want a simple user input. Every other widget is ignored if this is not null.
+  final Form userInputForm;
 
   route.FlushbarRoute<T> _flushbarRoute;
 
@@ -148,7 +232,8 @@ class Flushbar<T extends Object> extends StatefulWidget {
       flushbar: this,
     );
 
-    return await Navigator.of(context, rootNavigator: false).push(_flushbarRoute);
+    return await Navigator.of(context, rootNavigator: false)
+        .push(_flushbarRoute);
   }
 
   /// Dismisses the flushbar causing is to return a future containing [result].
@@ -188,31 +273,40 @@ class Flushbar<T extends Object> extends StatefulWidget {
   }
 }
 
-class _FlushbarState<K extends Object> extends State<Flushbar> with TickerProviderStateMixin {
-  FlushbarStatus currentStatus;
-
-  AnimationController _fadeController;
-  Animation<double> _fadeAnimation;
-
-  final Widget _emptyWidget = SizedBox(width: 0.0, height: 0.0);
+class _FlushbarState<K extends Object> extends State<Flushbar>
+    with TickerProviderStateMixin {
+  final Duration _pulseAnimationDuration = const Duration(seconds: 1);
+  final Widget _emptyWidget = const SizedBox();
   final double _initialOpacity = 1.0;
   final double _finalOpacity = 0.4;
 
-  final Duration _pulseAnimationDuration = Duration(seconds: 1);
-
+  GlobalKey _backgroundBoxKey;
+  FlushbarStatus currentStatus;
+  AnimationController _fadeController;
+  Animation<double> _fadeAnimation;
   bool _isTitlePresent;
   double _messageTopMargin;
-
+  FocusScopeNode _focusNode;
+  FocusAttachment _focusAttachment;
+  Completer<Size> _boxHeightCompleter;
+  Function _progressListener;
+  CurvedAnimation _progressAnimation;
 
   @override
   void initState() {
     super.initState();
 
-    assert(((widget.userInputForm != null || ((widget.message != null && widget.message.isNotEmpty) || widget.messageText != null))),
+    _backgroundBoxKey = GlobalKey();
+    _boxHeightCompleter = Completer<Size>();
+
+    assert(
+        ((widget.userInputForm != null ||
+            ((widget.message != null && widget.message.isNotEmpty) ||
+                widget.messageText != null))),
         "A message is mandatory if you are not using userInputForm. Set either a message or messageText");
 
     _isTitlePresent = (widget.title != null || widget.titleText != null);
-    _messageTopMargin = _isTitlePresent ? 6.0 : 16.0;
+    _messageTopMargin = _isTitlePresent ? 6.0 : widget.padding.top;
 
     _configureLeftBarFuture();
     _configureProgressIndicatorAnimation();
@@ -234,23 +328,30 @@ class _FlushbarState<K extends Object> extends State<Flushbar> with TickerProvid
     super.dispose();
   }
 
-  final Completer<double> _boxHeightCompleter = Completer<double>();
-
   void _configureLeftBarFuture() {
     SchedulerBinding.instance.addPostFrameCallback(
       (_) {
-        final keyContext = backgroundBoxKey.currentContext;
+        final keyContext = _backgroundBoxKey.currentContext;
 
         if (keyContext != null) {
           final RenderBox box = keyContext.findRenderObject();
-          _boxHeightCompleter.complete(box.size.height);
+          _boxHeightCompleter.complete(box.size);
         }
       },
     );
   }
 
+  void _configureProgressIndicatorAnimation() {
+    if (widget.showProgressIndicator &&
+        widget.progressIndicatorController != null) {
+      _progressAnimation = CurvedAnimation(
+          curve: Curves.linear, parent: widget.progressIndicatorController);
+    }
+  }
+
   void _configurePulseAnimation() {
-    _fadeController = AnimationController(vsync: this, duration: _pulseAnimationDuration);
+    _fadeController =
+        AnimationController(vsync: this, duration: _pulseAnimationDuration);
     _fadeAnimation = Tween(begin: _initialOpacity, end: _finalOpacity).animate(
       CurvedAnimation(
         parent: _fadeController,
@@ -270,28 +371,18 @@ class _FlushbarState<K extends Object> extends State<Flushbar> with TickerProvid
     _fadeController.forward();
   }
 
-  Function _progressListener;
-
-  void _configureProgressIndicatorAnimation() {
-    if (widget.showProgressIndicator && widget.progressIndicatorController != null) {
-      _progressListener = () {
-        setState(() {});
-      };
-      widget.progressIndicatorController.addListener(_progressListener);
-
-      _progressAnimation = CurvedAnimation(curve: Curves.linear, parent: widget.progressIndicatorController);
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
     return Align(
       heightFactor: 1.0,
       child: Material(
-        color: widget.flushbarStyle == FlushbarStyle.FLOATING ? Colors.transparent : widget.backgroundColor,
+        color: widget.flushbarStyle == FlushbarStyle.FLOATING
+            ? Colors.transparent
+            : widget.backgroundColor,
         child: SafeArea(
           minimum: widget.flushbarPosition == FlushbarPosition.BOTTOM
-              ? EdgeInsets.only(bottom: MediaQuery.of(context).viewInsets.bottom)
+              ? EdgeInsets.only(
+                  bottom: MediaQuery.of(context).viewInsets.bottom)
               : EdgeInsets.only(top: MediaQuery.of(context).viewInsets.top),
           bottom: widget.flushbarPosition == FlushbarPosition.BOTTOM,
           top: widget.flushbarPosition == FlushbarPosition.TOP,
@@ -304,38 +395,135 @@ class _FlushbarState<K extends Object> extends State<Flushbar> with TickerProvid
   }
 
   Widget _getFlushbar() {
-    return _generateFlushbar();
+    Widget flushbar;
+
+    if (widget.userInputForm != null) {
+      flushbar = _generateInputFlushbar();
+    } else {
+      flushbar = _generateFlushbar();
+    }
+
+    return Stack(
+      children: [
+        FutureBuilder(
+          future: _boxHeightCompleter.future,
+          builder: (context, AsyncSnapshot<Size> snapshot) {
+            if (snapshot.hasData) {
+              return ClipRect(
+                child: BackdropFilter(
+                  filter: ImageFilter.blur(
+                      sigmaX: widget.barBlur, sigmaY: widget.barBlur),
+                  child: Container(
+                    height: snapshot.data.height,
+                    width: snapshot.data.width,
+                    decoration: BoxDecoration(
+                      color: Colors.transparent,
+                      borderRadius: BorderRadius.circular(widget.borderRadius),
+                    ),
+                  ),
+                ),
+              );
+            } else {
+              return _emptyWidget;
+            }
+          },
+        ),
+        flushbar,
+      ],
+    );
   }
 
-  CurvedAnimation _progressAnimation;
-  GlobalKey backgroundBoxKey = GlobalKey();
-
-  Widget _generateFlushbar() {
-    return DecoratedBox(
-      key: backgroundBoxKey,
+  Widget _generateInputFlushbar() {
+    return Container(
+      key: _backgroundBoxKey,
+      constraints: widget.maxWidth != null
+          ? BoxConstraints(maxWidth: widget.maxWidth)
+          : null,
       decoration: BoxDecoration(
         color: widget.backgroundColor,
         gradient: widget.backgroundGradient,
         boxShadow: widget.boxShadows,
         borderRadius: BorderRadius.circular(widget.borderRadius),
+        border: widget.borderColor != null
+            ? Border.all(color: widget.borderColor, width: widget.borderWidth)
+            : null,
+      ),
+      child: Padding(
+        padding: const EdgeInsets.only(
+            left: 8.0, right: 8.0, bottom: 8.0, top: 16.0),
+        child: FocusScope(
+          child: widget.userInputForm,
+          node: _focusNode,
+          autofocus: true,
+        ),
+      ),
+    );
+  }
+
+  Widget _generateFlushbar() {
+    return Container(
+      key: _backgroundBoxKey,
+      constraints: widget.maxWidth != null
+          ? BoxConstraints(maxWidth: widget.maxWidth)
+          : null,
+      decoration: BoxDecoration(
+        color: widget.backgroundColor,
+        gradient: widget.backgroundGradient,
+        boxShadow: widget.boxShadows,
+        borderRadius: BorderRadius.circular(widget.borderRadius),
+        border: widget.borderColor != null
+            ? Border.all(color: widget.borderColor, width: widget.borderWidth)
+            : null,
       ),
       child: Column(
         mainAxisSize: MainAxisSize.min,
         children: [
-          widget.showProgressIndicator
-              ? LinearProgressIndicator(
-                  value: widget.progressIndicatorController != null ? _progressAnimation.value : null,
-                  backgroundColor: widget.progressIndicatorBackgroundColor,
-                  valueColor: widget.progressIndicatorValueColor,
-                )
-              : _emptyWidget,
-          Row(mainAxisSize: MainAxisSize.max, children: _getAppropriateRowLayout()),
+          _buildProgressIndicator(),
+          Row(
+            mainAxisSize: MainAxisSize.max,
+            children: _getAppropriateRowLayout(),
+          ),
         ],
       ),
     );
   }
 
+  Widget _buildProgressIndicator() {
+    if (widget.showProgressIndicator && _progressAnimation != null) {
+      return AnimatedBuilder(
+          animation: _progressAnimation,
+          builder: (_, __) {
+            return LinearProgressIndicator(
+              value: _progressAnimation.value,
+              backgroundColor: widget.progressIndicatorBackgroundColor,
+              valueColor: widget.progressIndicatorValueColor,
+            );
+          });
+    }
+
+    if (widget.showProgressIndicator) {
+      return LinearProgressIndicator(
+        backgroundColor: widget.progressIndicatorBackgroundColor,
+        valueColor: widget.progressIndicatorValueColor,
+      );
+    }
+
+    return _emptyWidget;
+  }
+
   List<Widget> _getAppropriateRowLayout() {
+    double buttonRightPadding;
+    double iconPadding = 0;
+    if (widget.padding.right - 12 < 0) {
+      buttonRightPadding = 4;
+    } else {
+      buttonRightPadding = widget.padding.right - 12;
+    }
+
+    if (widget.padding.left > 16.0) {
+      iconPadding = widget.padding.left;
+    }
+
     if (widget.icon == null && widget.mainButton == null) {
       return [
         _buildLeftBarIndicator(),
@@ -347,12 +535,21 @@ class _FlushbarState<K extends Object> extends State<Flushbar> with TickerProvid
             children: <Widget>[
               (_isTitlePresent)
                   ? Padding(
-                      padding: const EdgeInsets.only(top: 16.0, left: 16.0, right: 16.0),
+                      padding: EdgeInsets.only(
+                        top: widget.padding.top,
+                        left: widget.padding.left,
+                        right: widget.padding.right,
+                      ),
                       child: _getTitleText(),
                     )
                   : _emptyWidget,
               Padding(
-                padding: EdgeInsets.only(top: _messageTopMargin, left: 16.0, right: 16.0, bottom: 16.0),
+                padding: EdgeInsets.only(
+                  top: _messageTopMargin,
+                  left: widget.padding.left,
+                  right: widget.padding.right,
+                  bottom: widget.padding.bottom,
+                ),
                 child: widget.messageText ?? _getDefaultNotificationText(),
               ),
             ],
@@ -363,7 +560,7 @@ class _FlushbarState<K extends Object> extends State<Flushbar> with TickerProvid
       return <Widget>[
         _buildLeftBarIndicator(),
         ConstrainedBox(
-          constraints: BoxConstraints.tightFor(width: 42.0),
+          constraints: BoxConstraints.tightFor(width: 42.0 + iconPadding),
           child: _getIcon(),
         ),
         Expanded(
@@ -374,12 +571,21 @@ class _FlushbarState<K extends Object> extends State<Flushbar> with TickerProvid
             children: <Widget>[
               (_isTitlePresent)
                   ? Padding(
-                      padding: const EdgeInsets.only(top: 16.0, left: 4.0, right: 16.0),
+                      padding: EdgeInsets.only(
+                        top: widget.padding.top,
+                        left: 4.0,
+                        right: widget.padding.left,
+                      ),
                       child: _getTitleText(),
                     )
                   : _emptyWidget,
               Padding(
-                padding: EdgeInsets.only(top: _messageTopMargin, left: 4.0, right: 16.0, bottom: 16.0),
+                padding: EdgeInsets.only(
+                  top: _messageTopMargin,
+                  left: 4.0,
+                  right: widget.padding.right,
+                  bottom: widget.padding.bottom,
+                ),
                 child: widget.messageText ?? _getDefaultNotificationText(),
               ),
             ],
@@ -397,26 +603,38 @@ class _FlushbarState<K extends Object> extends State<Flushbar> with TickerProvid
             children: <Widget>[
               (_isTitlePresent)
                   ? Padding(
-                      padding: const EdgeInsets.only(top: 16.0, left: 16.0, right: 16.0),
+                      padding: EdgeInsets.only(
+                        top: widget.padding.top,
+                        left: widget.padding.left,
+                        right: widget.padding.right,
+                      ),
                       child: _getTitleText(),
                     )
                   : _emptyWidget,
               Padding(
-                padding: EdgeInsets.only(top: _messageTopMargin, left: 16.0, right: 8.0, bottom: 16.0),
+                padding: EdgeInsets.only(
+                  top: _messageTopMargin,
+                  left: widget.padding.left,
+                  right: 8.0,
+                  bottom: widget.padding.bottom,
+                ),
                 child: widget.messageText ?? _getDefaultNotificationText(),
               ),
             ],
           ),
         ),
         Padding(
-          padding: const EdgeInsets.only(right: 4.0),
+          padding: EdgeInsets.only(right: buttonRightPadding),
           child: _getMainActionButton(),
         ),
       ];
     } else {
       return <Widget>[
         _buildLeftBarIndicator(),
-        ConstrainedBox(constraints: BoxConstraints.tightFor(width: 42.0), child: _getIcon()),
+        ConstrainedBox(
+          constraints: BoxConstraints.tightFor(width: 42.0 + iconPadding),
+          child: _getIcon(),
+        ),
         Expanded(
           flex: 1,
           child: Column(
@@ -425,19 +643,28 @@ class _FlushbarState<K extends Object> extends State<Flushbar> with TickerProvid
             children: <Widget>[
               (_isTitlePresent)
                   ? Padding(
-                      padding: const EdgeInsets.only(top: 16.0, left: 4.0, right: 8.0),
+                      padding: EdgeInsets.only(
+                        top: widget.padding.top,
+                        left: 4.0,
+                        right: 8.0,
+                      ),
                       child: _getTitleText(),
                     )
                   : _emptyWidget,
               Padding(
-                padding: EdgeInsets.only(top: _messageTopMargin, left: 4.0, right: 8.0, bottom: 16.0),
+                padding: EdgeInsets.only(
+                  top: _messageTopMargin,
+                  left: 4.0,
+                  right: 8.0,
+                  bottom: widget.padding.bottom,
+                ),
                 child: widget.messageText ?? _getDefaultNotificationText(),
               ),
             ],
           ),
         ),
         Padding(
-              padding: const EdgeInsets.only(right: 4.0),
+              padding: EdgeInsets.only(right: buttonRightPadding),
               child: _getMainActionButton(),
             ) ??
             _emptyWidget,
@@ -449,12 +676,12 @@ class _FlushbarState<K extends Object> extends State<Flushbar> with TickerProvid
     if (widget.leftBarIndicatorColor != null) {
       return FutureBuilder(
         future: _boxHeightCompleter.future,
-        builder: (BuildContext buildContext, AsyncSnapshot<double> snapshot) {
+        builder: (BuildContext buildContext, AsyncSnapshot<Size> snapshot) {
           if (snapshot.hasData) {
             return Container(
               color: widget.leftBarIndicatorColor,
               width: 5.0,
-              height: snapshot.data,
+              height: snapshot.data.height,
             );
           } else {
             return _emptyWidget;
@@ -484,7 +711,10 @@ class _FlushbarState<K extends Object> extends State<Flushbar> with TickerProvid
         ? widget.titleText
         : Text(
             widget.title ?? "",
-            style: TextStyle(fontSize: 16.0, color: Colors.white, fontWeight: FontWeight.bold),
+            style: TextStyle(
+                fontSize: 16.0,
+                color: Colors.white,
+                fontWeight: FontWeight.bold),
           );
   }
 
